requirements.txt
fastapi~=0.111.0
pydantic~=2.7.4
requests~=2.32.3
python-dotenv~=1.0.1